\name{aCGH.Spline-package}
\alias{aCGH.Spline}
\docType{package}
\title{
Package for robust spline interpolation normalisation on dual color aCGH data. 
}
\description{
Carries out a robust spline interpolation method on aCGH dual color microarray data.\cr\cr
Contains some functions for reading and writing of standard aCGH data formats along with the generation QC statistics.\cr
}
\details{
\tabular{ll}{
Package: \tab aCGH.Spline\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-05-14\cr
License: \tab Open Source\cr
LazyLoad: \tab yes\cr
}

}
\author{
Tomas William Fitzgerald

Maintainer: tf2@sanger.ac.uk
}

\keyword{ package }
